fixes and formatting: - add timestamps to folders and tags - replace passport with custom middleware - add mongoose utils - update seed data for variety - add lots of test coverage

+++ routes/folders.js
@@ -1,42 +1,39 @@
-'use strict';
+"use strict";
 
-const express = require('express');
-const mongoose = require('mongoose');
-const passport = require('passport');
+const express = require("express");
+const mongoose = require("mongoose");
 
-const Folder = require('../models/folder');
-const Note = require('../models/note');
+const Folder = require("../models/folder");
+const Note = require("../models/note");
 
 const router = express.Router();
 
-// Protect endpoints using JWT Strategy
-router.use('/folders', passport.authenticate('jwt', { session: false, failWithError: true }));
-
 /* ========== GET/READ ALL ITEMS ========== */
-router.get('/folders', (req, res, next) => {
+router.get("/", (req, res, next) => {
+  const userId = req.user.id;
 
-  Folder.find({ userId: req.user.id })
-    .sort('name')
+  Folder.find({ userId })
+    .sort("name")
     .then(results => {
       res.json(results);
     })
     .catch(err => {
       next(err);
     });
-
 });
 
 /* ========== GET/READ A SINGLE ITEM ========== */
-router.get('/folders/:id', (req, res, next) => {
+router.get("/:id", (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    const err = new Error('The `id` is not valid');
+    const err = new Error("The `id` is not valid");
     err.status = 400;
     return next(err);
   }
 
-  Folder.findOne({ _id: id, userId: req.user.id })
+  Folder.findOne({ _id: id, userId })
     .then(result => {
       if (result) {
         res.json(result);
@@ -47,30 +44,29 @@ router.get('/folders/:id', (req, res, next) => {
     .catch(err => {
       next(err);
     });
-
 });
 
 /* ========== POST/CREATE AN ITEM ========== */
-router.post('/folders', (req, res, next) => {
+router.post("/", (req, res, next) => {
   const { name } = req.body;
   const userId = req.user.id;
 
-  const newItem = { name, userId };
+  const newFolder = { name, userId };
 
   /***** Never trust users - validate input *****/
   if (!name) {
-    const err = new Error('Missing `name` in request body');
+    const err = new Error("Missing `name` in request body");
     err.status = 400;
     return next(err);
   }
 
-  Folder.create(newItem)
+  Folder.create(newFolder)
     .then(result => {
       res.location(`${req.originalUrl}/${result.id}`).status(201).json(result);
     })
     .catch(err => {
       if (err.code === 11000) {
-        err = new Error('The folder name already exists');
+        err = new Error("Folder name already exists");
         err.status = 400;
       }
       next(err);
@@ -78,27 +74,27 @@ router.post('/folders', (req, res, next) => {
 });
 
 /* ========== PUT/UPDATE A SINGLE ITEM ========== */
-router.put('/folders/:id', (req, res, next) => {
+router.put("/:id", (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
   const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
-  if (!name) {
-    const err = new Error('Missing `name` in request body');
-    err.status = 400;
-    return next(err);
-  }
-
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    const err = new Error('The `id` is not valid');
+    const err = new Error("The `id` is not valid");
     err.status = 400;
     return next(err);
   }
 
-  const updateItem = { name, userId };
+  if (!name) {
+    const err = new Error("Missing `name` in request body");
+    err.status = 400;
+    return next(err);
+  }
 
-  Folder.findOneAndUpdate({ _id: id, userId }, updateItem, { new: true })
+  const updateFolder = { name, userId };
+
+  Folder.findByIdAndUpdate(id, updateFolder, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -108,7 +104,7 @@ router.put('/folders/:id', (req, res, next) => {
     })
     .catch(err => {
       if (err.code === 11000) {
-        err = new Error('The folder name already exists');
+        err = new Error("Folder name already exists");
         err.status = 400;
       }
       next(err);
@@ -116,15 +112,22 @@ router.put('/folders/:id', (req, res, next) => {
 });
 
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
-router.delete('/folders/:id', (req, res, next) => {
+router.delete("/:id", (req, res, next) => {
   const { id } = req.params;
   const userId = req.user.id;
+
+  /***** Never trust users - validate input *****/
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
 
   const folderRemovePromise = Folder.findOneAndRemove({ _id: id, userId });
 
   const noteRemovePromise = Note.updateMany(
     { folderId: id, userId },
-    { $unset: { folderId: '' } }
+    { $unset: { folderId: "" } }
   );
 
   Promise.all([folderRemovePromise, noteRemovePromise])
