+++ routes/notes.js
@@ -1,67 +1,75 @@
-'use strict';
+"use strict";
 
-const express = require('express');
-const mongoose = require('mongoose');
-const passport = require('passport');
+const express = require("express");
+const mongoose = require("mongoose");
 
-const Note = require('../models/note');
-const Folder = require('../models/folder');
-const Tag = require('../models/tag');
+const Note = require("../models/note");
+const Folder = require("../models/folder");
+const Tag = require("../models/tag");
 
 const router = express.Router();
 
-// Protect endpoints using JWT Strategy
-router.use('/notes', passport.authenticate('jwt', { session: false, failWithError: true }));
-
-function validateFolderId(userId, folderId) {
-  if (!folderId) {
+function validateFolderId(folderId, userId) {
+  if (folderId === undefined) {
     return Promise.resolve();
   }
-  return Folder.findOne({ _id: folderId, userId })
-    .then(result => {
-      if (!result) {
-        return Promise.reject('InvalidFolder');
+  if (!mongoose.Types.ObjectId.isValid(folderId)) {
+    const err = new Error("The `folderId` is not valid");
+    err.status = 400;
+    return Promise.reject(err);
+  }
+  return Folder.count({ _id: folderId, userId })
+    .then(count => {
+      if (count === 0) {
+        const err = new Error("The `folderId` is not valid");
+        err.status = 400;
+        return Promise.reject(err);
       }
     });
 }
 
-function validateTagIds(userId, tags = []) {
-  if (!tags.length) {
+function validateTagIds(tags, userId) {
+  if (tags === undefined) {
     return Promise.resolve();
+  }
+  if (!Array.isArray(tags)) {
+    const err = new Error("The `tags` must be an array");
+    err.status = 400;
+    return Promise.reject(err);
   }
   return Tag.find({ $and: [{ _id: { $in: tags }, userId }] })
     .then(results => {
       if (tags.length !== results.length) {
-        return Promise.reject('InvalidTag');
+        const err = new Error("The `tags` contains an non existent id");
+        err.status = 400;
+        return Promise.reject(err);
       }
     });
 }
 
 /* ========== GET/READ ALL ITEMS ========== */
-router.get('/notes', (req, res, next) => {
+router.get("/", (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
   const userId = req.user.id;
 
   let filter = { userId };
-  let sort = 'created'; // default sorting
 
-  // if querying by searchTerm, then add to filter
   if (searchTerm) {
-    const re = new RegExp(searchTerm, 'i');
-    filter.$or = [{ 'title': { $regex: re } }, { 'content': { $regex: re } }];
+    // filter.title = { $regex: searchTerm };
+    filter.$or = [{ "title": { $regex: searchTerm } }, { "content": { $regex: searchTerm } }];
   }
 
-  // if querying by folder, then add to filter
   if (folderId) {
     filter.folderId = folderId;
   }
 
-  // if querying by tags, then add to filter
   if (tagId) {
     filter.tags = tagId;
   }
 
-  Note.find(filter).populate('tags').sort(sort)
+  Note.find(filter)
+    .populate("tags")
+    .sort({ "updatedAt": "desc" })
     .then(results => {
       res.json(results);
     })
@@ -71,18 +79,19 @@ router.get('/notes', (req, res, next) => {
 });
 
 /* ========== GET/READ A SINGLE ITEM ========== */
-router.get('/notes/:id', (req, res, next) => {
+router.get("/:id", (req, res, next) => {
   const { id } = req.params;
   const userId = req.user.id;
 
+  /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    const err = new Error('The `id` is not valid');
+    const err = new Error("The `id` is not valid");
     err.status = 400;
     return next(err);
   }
 
   Note.findOne({ _id: id, userId })
-    .populate('tags')
+    .populate("tags")
     .then(result => {
       if (result) {
         res.json(result);
@@ -96,78 +105,71 @@ router.get('/notes/:id', (req, res, next) => {
 });
 
 /* ========== POST/CREATE AN ITEM ========== */
-router.post('/notes', (req, res, next) => {
+router.post("/", (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
   const userId = req.user.id;
+  const newNote = { title, content, userId, folderId, tags };
 
   /***** Never trust users - validate input *****/
   if (!title) {
-    const err = new Error('Missing `title` in request body');
+    const err = new Error("Missing `title` in request body");
     err.status = 400;
     return next(err);
   }
 
-  const newItem = { title, content, folderId, tags, userId };
-  const valFolderIdProm = validateFolderId(userId, folderId);
-  const valTagIdsProm = validateTagIds(userId, tags);
-
-  Promise.all([valFolderIdProm, valTagIdsProm])
-    .then(() => Note.create(newItem))
+  Promise.all([
+    validateFolderId(folderId, userId),
+    validateTagIds(tags, userId)
+  ])
+    .then(() => Note.create(newNote))
     .then(result => {
       res.location(`${req.originalUrl}/${result.id}`).status(201).json(result);
     })
     .catch(err => {
-      if (err === 'InvalidFolder') {
-        err = new Error('The folder is not valid');
+      if (err === "InvalidFolder") {
+        err = new Error("The folder is not valid");
         err.status = 400;
       }
-      if (err === 'InvalidTag') {
-        err = new Error('The tag is not valid');
+      if (err === "InvalidTag") {
+        err = new Error("The tag is not valid");
         err.status = 400;
       }
       next(err);
     });
-
-  // Note.create(newItem)
-  // .then(result => {
-  //   res.location(`${req.originalUrl}/${result.id}`).status(201).json(result);
-  // })
-  // .catch(err => {       next(err);     });
 });
 
 /* ========== PUT/UPDATE A SINGLE ITEM ========== */
-router.put('/notes/:id', (req, res, next) => {
+router.put("/:id", (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags } = req.body;
   const userId = req.user.id;
+  const updateNote = { title, content, userId, folderId, tags };
 
   /***** Never trust users - validate input *****/
-  if (!title) {
-    const err = new Error('Missing `title` in request body');
-    err.status = 400;
-    return next(err);
-  }
-
   if (!mongoose.Types.ObjectId.isValid(id)) {
-    const err = new Error('The `id` is not valid');
+    const err = new Error("The `id` is not valid");
     err.status = 400;
     return next(err);
   }
 
-  const updateItem = { title, content, tags, userId, folderId };
+  if (!title) {
+    const err = new Error("Missing `title` in request body");
+    err.status = 400;
+    return next(err);
+  }
 
   if (mongoose.Types.ObjectId.isValid(folderId)) {
-    updateItem.folderId = folderId;
+    updateNote.folderId = folderId;
   }
 
-  const options = { new: true };
-
-  const valFolderIdProm = validateFolderId(userId, folderId);
-  const valTagIdsProm = validateTagIds(userId, tags);
-
-  Promise.all([valFolderIdProm, valTagIdsProm])
-    .then(() => Note.findByIdAndUpdate(id, updateItem, options).populate('tags')
-    )
+  Promise.all([
+    validateFolderId(folderId, userId),
+    validateTagIds(tags, userId)
+  ])
+    .then(() => {
+      return Note.findByIdAndUpdate(id, updateNote, { new: true })
+        .populate("tags");
+    })
     .then(result => {
       if (result) {
         res.json(result);
@@ -176,23 +178,29 @@ router.put('/notes/:id', (req, res, next) => {
       }
     })
     .catch(err => {
-      if (err === 'InvalidFolder') {
-        err = new Error('The folder is not valid');
+      if (err === "InvalidFolder") {
+        err = new Error("The folder is not valid");
         err.status = 400;
       }
-      if (err === 'InvalidTag') {
-        err = new Error('The tag is not valid');
+      if (err === "InvalidTag") {
+        err = new Error("The tag is not valid");
         err.status = 400;
       }
       next(err);
     });
-
 });
 
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
-router.delete('/notes/:id', (req, res, next) => {
+router.delete("/:id", (req, res, next) => {
   const { id } = req.params;
   const userId = req.user.id;
+
+  /***** Never trust users - validate input *****/
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error("The `id` is not valid");
+    err.status = 400;
+    return next(err);
+  }
 
   Note.findOneAndRemove({ _id: id, userId })
     .then(() => {
